Reworked Invoicespec
Changed some fields
Added better (and more conistent with the rest of ON-api docs) field specifications

Added new dataformat number due to invoice price with decimals.

Added connectionDate and disconnectionDate to Subscriptions endpoint as that information might be useful. The data is in the invoice spec anyway so it is needed there.

Updated some field names to match field names i access and subscription.